+++ poulpy-core/src/tests/test_suite/mod.rs
@@ -1,6 +1,7 @@
 pub mod automorphism;
 pub mod encryption;
 pub mod external_product;
+pub mod glwe_tensor;
 pub mod keyswitch;
 
 mod conversion;